Update /pushers/set info to make deletion clearer
I had trouble figuring out how to remove a pusher with this endpoint today, and had to look synapse's implementation to figure it out. Later @richvdh  pointed out that this bit of info was indeed already there, in the `kind` property: "null deletes the pusher.", but I feel it would still be valuable to add this info to the top and specify what additional params required for the operation (at least according to synapse). 
Maybe the text I added needs some changes to make it more align with the pattern from the rest of the specs, let me know.