Add a hyphen between third and party when used as an adjective
'third-party' should have a hyphen when used as an adjective, such as in
'third-party network'. It should not have a dash when used as a noun, such
as in 'sharing with a third party'.

All uses in the spec of third-party were as an adjective. Thus they have all
been updated to include a hyphen where they didn't already.