Pass `additionalProperties` and `patternProperties` into `render-object-table`
Previously, we were stripping `additionalProperties` and `patternProperties`
from all objects except top-level objects. Obviously, this was no good for
objects where a nested property contains such properties.

Fixing that (in `clean-object`) *ought* to be simple enough, except that it
turned out we were relying on the fact that would give us an "empty" entry in
the array of types-to-render-tables-for returned by
`resolve-additional-types`. (Normally, we don't want an object that only has
`additionalProperties` to have its own table, since we just embed it in the
parent table.)

So, we need to add more logic to `resolve-additional-types-inner` to suppress
such tables.

This commit doesn't change the rendered output at all (verified via `diff`):
it's just preparation for what comes next.