Remove `withVersioning` parameter of `rver-fragment` shortcode.
We now detect automatically in the `added-in` and `changed-in` shortcodes
whether we need to add text according to the page's room version so this parameter
has become unnecessary.
Moreover, forgetting to set this param can remove the text generated by the
shortcodes so it is safer to just get rid of it.

Signed-off-by: Kévin Commaille <zecakeh@tedomum.fr>