Place the endpoints list just under the description of each API
We needed to override Docsy's default page renderers to show it. Because
the Client-Server API uses a different render chain than the other
endpoint pages, we have to override two of Docsy's pages (`content.html` and `single.html`).

In order to reduce duplication, we then put the shared content into a
`spec-content.html` partial.